Replace old reader with reader2
