+++ package.json
@@ -18,8 +18,8 @@
 		"@melt-ui/pp": "^0.1.2",
 		"@playwright/test": "^1.28.1",
 		"@sveltejs/adapter-auto": "^2.0.0",
-		"@sveltejs/adapter-static": "^2.0.2",
-		"@sveltejs/kit": "^1.24.1",
+		"@sveltejs/adapter-static": "^2.0.3",
+		"@sveltejs/kit": "^1.25.1",
 		"@tauri-apps/cli": "^1.4.0",
 		"@types/postcss-prefix-selector": "^1.15.0",
 		"@typescript-eslint/eslint-plugin": "^5.45.0",
@@ -31,8 +31,8 @@
 		"postcss": "^8.4.26",
 		"prettier": "^2.8.0",
 		"prettier-plugin-svelte": "^2.10.1",
-		"svelte": "^4.0.5",
-		"svelte-check": "^3.5.1",
+		"svelte": "^4.2.1",
+		"svelte-check": "^3.5.2",
 		"svelte-sequential-preprocessor": "^2.0.1",
 		"tailwindcss": "^3.3.3",
 		"tslib": "^2.4.1",
